+++ bin/reth/src/node/mod.rs
@@ -11,7 +11,7 @@ use crate::{
 use clap::{crate_version, Parser};
 use eyre::Context;
 use fdlimit::raise_fd_limit;
-use futures::{future::Either, pin_mut, stream::select as stream_select, StreamExt};
+use futures::{pin_mut, stream::select as stream_select, StreamExt};
 use reth_auto_seal_consensus::{AutoSealBuilder, AutoSealConsensus};
 use reth_basic_payload_builder::{BasicPayloadJobGenerator, BasicPayloadJobGeneratorConfig};
 use reth_beacon_consensus::{BeaconConsensus, BeaconConsensusEngine, BeaconEngineMessage};
@@ -33,7 +33,7 @@ use reth_interfaces::{
     consensus::{Consensus, ForkchoiceState},
     p2p::{
         bodies::{client::BodiesClient, downloader::BodyDownloader},
-        download::DownloadClient,
+        either::EitherDownloader,
         headers::{client::StatusUpdater, downloader::HeaderDownloader},
     },
     sync::SyncStateUpdater,
@@ -705,80 +705,6 @@ impl Command {
             .build(db);
 
         Ok(pipeline)
-    }
-}
-
-/// A sum type to make initializing a consensus engine with different client types easier.
-#[derive(Debug, Clone)]
-pub enum EitherDownloader<A, B> {
-    /// The first downloader variant
-    Left(A),
-    /// The second downloader variant
-    Right(B),
-}
-
-impl<A, B> DownloadClient for EitherDownloader<A, B>
-where
-    A: DownloadClient,
-    B: DownloadClient,
-{
-    fn report_bad_message(&self, peer_id: reth_primitives::PeerId) {
-        match self {
-            EitherDownloader::Left(a) => a.report_bad_message(peer_id),
-            EitherDownloader::Right(b) => b.report_bad_message(peer_id),
-        }
-    }
-    fn num_connected_peers(&self) -> usize {
-        match self {
-            EitherDownloader::Left(a) => a.num_connected_peers(),
-            EitherDownloader::Right(b) => b.num_connected_peers(),
-        }
-    }
-}
-
-impl<A, B> BodiesClient for EitherDownloader<A, B>
-where
-    A: BodiesClient,
-    B: BodiesClient,
-{
-    type Output = Either<A::Output, B::Output>;
-
-    fn get_block_bodies_with_priority(
-        &self,
-        hashes: Vec<H256>,
-        priority: reth_interfaces::p2p::priority::Priority,
-    ) -> Self::Output {
-        match self {
-            EitherDownloader::Left(a) => {
-                Either::Left(a.get_block_bodies_with_priority(hashes, priority))
-            }
-            EitherDownloader::Right(b) => {
-                Either::Right(b.get_block_bodies_with_priority(hashes, priority))
-            }
-        }
-    }
-}
-
-impl<A, B> HeadersClient for EitherDownloader<A, B>
-where
-    A: HeadersClient,
-    B: HeadersClient,
-{
-    type Output = Either<A::Output, B::Output>;
-
-    fn get_headers_with_priority(
-        &self,
-        request: reth_interfaces::p2p::headers::client::HeadersRequest,
-        priority: reth_interfaces::p2p::priority::Priority,
-    ) -> Self::Output {
-        match self {
-            EitherDownloader::Left(a) => {
-                Either::Left(a.get_headers_with_priority(request, priority))
-            }
-            EitherDownloader::Right(b) => {
-                Either::Right(b.get_headers_with_priority(request, priority))
-            }
-        }
     }
 }
